requirements.txt
matplotlib
numpy
opencv-python
Pillow
roboflow
scikit-learn
seaborn
torch
torchmetrics
torchvision
tqdm
ultralytics
